test(devices): split device_sync test harness to stay under the 300-line limit
Extract the shared mock harness into harness.mjs and the provider tests into
intune.test.mjs / jumpcloud.test.mjs; test.mjs is now a thin runner. It imports
the two suites sequentially via dynamic import because they use top-level await
and stub the global fetch — static imports would interleave and clobber it.
